requirements.txt
langchain
langchain-community
pymupdf
tiktoken
python-dotenv
langchain-openai
langchain-google-genai
duckduckgo-search